Externalize zoomhome plugin and set options Move addControlPlaceholders function up
